use transactions for S-mode messages
The transactions inside of the stack were designed to keep track of
which CON message you need to send an acknowledgement to. NONs were
completely excluded from this mechanism.

This commit makes S-mode messages use transactions in order for the
retransmission of the packet with the included Echo request to work.
This is still work in progress - the NON transactions aren't freed in
all cases so there is likely a memory leak.

The first message after reboot doesn't work, but past that it's smooth
sailing. Also, there is an unexpectedly large amount of communication on
the wire - the ACKs and RSTs should not really happen when using NON
communication.